refacto context manager
