+++ DisplayBundle/DisplayBlock/Strategies/FooterStrategy.php
@@ -88,7 +88,7 @@ class FooterStrategy extends AbstractAuthorizationCheckerStrategy
      */
     protected function getNodes()
     {
-        $language = $this->currentSiteManager->getCurrentSiteDefaultLanguage();
+        $language = $this->currentSiteManager->getCurrentSiteLanguage();
         $siteId = $this->currentSiteManager->getCurrentSiteId();
         $nodes = $this->nodeRepository->getFooterTree($language, $siteId);
