+++ DisplayBundle/Manager/NodeManager.php
@@ -6,7 +6,6 @@ use OpenOrchestra\DisplayBundle\Exception\NodeNotFoundException;
 use OpenOrchestra\ModelInterface\Model\ReadNodeInterface;
 use OpenOrchestra\ModelInterface\Repository\ReadNodeRepositoryInterface;
 use OpenOrchestra\ModelInterface\Repository\ReadSiteRepositoryInterface;
-use OpenOrchestra\BaseBundle\Context\CurrentSiteIdInterface;
 
 /**
  * Class NodeManager
@@ -20,14 +19,13 @@ class NodeManager
     /**
      * @param ReadNodeRepositoryInterface $nodeRepository
      * @param ReadSiteRepositoryInterface $siteRepository
-     * @param CurrentSiteIdInterface      $currentSiteManager
+     * @param ContextInterface            $currentSiteManager
      */
     public function __construct(
         ReadNodeRepositoryInterface $nodeRepository,
         ReadSiteRepositoryInterface $siteRepository,
-        CurrentSiteIdInterface      $currentSiteManager
-    )
-    {
+        ContextInterface            $currentSiteManager
+    ) {
         $this->nodeRepository = $nodeRepository;
         $this->siteRepository = $siteRepository;
         $this->currentSiteManager = $currentSiteManager;
